+++ doc/ChangeLog.md
@@ -9,6 +9,8 @@
 - BoolArray clear, eachTrue, getAndSet
 - Int.clip, Float.clip
 - Bit I/O: OutStream.writeBits, InStream.readBits
+- Allow once methods to be used on const classes
+- Add Closure flags
 
 ### Build 3.3 (2020-10-25)
 - Fix NioBuf